utils/requirements.txt
pandas==2.1.4
numpy==1.26.4
networkx==3.4.2
scipy==1.15.0
SQLAlchemy==1.4.54
python-dotenv==1.0.1
fireducks==1.1.6
requests==2.32.3
hdbscan==0.8.40
scikit-learn==1.6.0
pyproj==3.7.0
neo4j==5.27.0
psycopg2-binary==2.9.10
geopandas==1.0.1
pyogrio==0.10.0
kagglehub==0.3.6
shapely==2.0.6
dateutils==0.6.12
redis==5.2.1